ignite-6273 Moved ServiceDeploymentException to org.apache.ignite.services
